requirements.txt
# Core dependencies for business process simulation
pandas>=2.0.0
psycopg2-binary>=2.9.0
numpy>=1.24.0

# Data fetching from Kaggle
mlcroissant>=0.0.1

# Data processing and analysis
openpyxl>=3.1.0
xlrd>=2.0.0

# Logging and configuration
python-dotenv>=1.0.0

# Optional: For enhanced data visualization and reporting
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For advanced data processing
scikit-learn>=1.3.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Database utilities
sqlalchemy>=2.0.0
alembic>=1.11.0
